+++ .gitignore
@@ -1,1 +1,4 @@
-node_modules
+/node_modules
+/coverage.data
+/coverage/
+/dist
