--- .travis.yml
@@ -1,6 +1,0 @@
-language: node_js
-node_js:
-  - 12
-  - 11
-  - 10
-  - 8
