+++ package.json
@@ -2,11 +2,17 @@
   "name": "mongo-date-query",
   "version": "1.3.1",
   "description": "Mongo date query helper for common queries",
-  "main": "index.js",
+  "main": "./dist/index.js",
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.ts",
   "repository": "TalhaAwan/mongo-date-query",
-  "homepage": "https://github.com/TalhaAwan/mongo-date-query#readme",
+  "files": [
+    "dist"
+  ],
   "scripts": {
-    "test": "xo && ava"
+    "build": "tsup",
+    "test": "jest",
+    "prepare": "npm run build"
   },
   "keywords": [
     "mongo",
@@ -18,11 +24,15 @@
   "author": "Talha Awan",
   "license": "MIT",
   "devDependencies": {
-    "ava": "^2.4.0",
-    "xo": "^0.25.4",
-    "mockdate": "^3.0.2"
+    "mockdate": "^3.0.5"
   },
   "dependencies": {
-    "moment": "^2.29.1"
+    "@types/jest": "^29.5.12",
+    "jest": "^29.7.0",
+    "moment": "^2.30.1",
+    "ts-jest": "^29.2.5",
+    "ts-node": "^10.9.2",
+    "tsup": "^8.2.4",
+    "typescript": "^5.6.2"
   }
 }
